requirements.txt.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
matplotlib==3.9.0
gunicorn==21.2.0
python-dotenv==1.0.1
openai>=1.0.0

Werkzeug==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
SQLAlchemy==2.0.34